+++ src/App.tsx
@@ -1,6 +1,6 @@
-import AuthProtected from 'components/common/AuthProtected';
-import Layout from 'components/common/Layout';
-import RedirectToTodoPageIfLoggedIn from 'components/common/RedirectToTodoPageIfLoggedIn';
+import AuthProtected from 'components/routes/AuthProtected';
+import Layout from 'components/routes/Layout';
+import RedirectToTodoPageIfLoggedIn from 'components/routes/RedirectToTodoPageIfLoggedIn';
 import NotFound from 'pages/NotFound';
 import Signin from 'pages/Signin';
 import Signup from 'pages/Signup';
